requirements.txt
fastapi==0.115.12
pydantic==2.11.4
torch==2.6.0
transformers==4.51.3
nest-asyncio==1.6.0
pyngrok==7.2.7
uvicorn==0.34.2
python-dotenv==1.0.0
accelerate==1.6.0